hubble: move some types definition to avoid import cycles
[ upstream commit e54c0008ab40dd6c40b10e44d659774cd8f67ef5 ]

As part of implementing unit tests for `pkg/hubble/relay/observer`,
`FakeClientConn` is needed. Prior to this commit, the implementation was
only present for testing in `pkg/hubble/relay/pool`. However, moving it
to `pkg/hubble/testutils` creates import cycles within unit test files.
In order to resolve them, some types definitions are moved to the
existing `pkg/hubble/peer/types` package and a new
`pkg/hubble/relay/pool/types` package.

There are no functional changes with this commit; only the necessary
changes to avoid build/test failures resulting from the aforementioned
reorganization of packages and types definitions.
